refactor(calendar-view): move the select hour logic into a new hook
